Support for more than 3 numbers in C# number relations challenge
Now the program allows (theoretically) infinite number... of numbers. As
long as you've got the processing power and memory. :-P

Changed the operations to work on double-precision numbers instead of
long integers to allow decimal numbers.

Also, out of curiosity, added modulo operator and it seems to work fine.
Left it disabled as it gluttered the output too much.